+++ src/main/java/vkhanhqui/myblog/models/repositories/TagRepositories.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import vkhanhqui.myblog.models.Tag;
 
 @Repository("tagRepository")
-public interface TagRepositories extends JpaRepository<Tag, Integer> {
+public interface TagRepositories extends JpaRepository<Tag, Long> {
 }
